+++ examples/mixed/run.py
@@ -1,4 +1,3 @@
-import numpy as np
 import argparse
 import matplotlib.pyplot as plt